set rbldnsd refresh to 0 (disabled) and implemented SIGHUP
expurgate-solo will now only tell rbldnsd using funciton rbldnsrefresh() which sends a SIGHUP  to rbldnsd to look for changes in the /running-config file if there is a change detected and not ever second per default rbldsnd behaviour. Plan to add and environment variable so this will not occur if rbldnsd -c != 0

restdb improvements to add resiliency if DNS is down, or  'no internet connection' or it is not possible to connect to restdb.

For records that count toward dnscount, errors are counted and if more than 0 - no config files will written. This may need to be extended to NXDOMAIN where a relay is in between expurgate and the upstream DNS forwarders